adjusted how much shots are curved (increased it) based on user feedback.
Increased the duration that Z and X are pressed in driving range mode to help key presses not get missed (after a bunch of testing I haven't seen a missed key press since, so hopefully it's good now).

turned off club_lockstep at the start of takeShot and return it to it's previous state at the end.  This prevents the club from changing names in the GUI durring a shot (wouldn't affect the shot, but looks bad), and requires less processing which may be beneficial.